\name{HierO-package}
\alias{HierO-package}
\docType{package}
\title{GUI for optimizing hierarchical data structure}

\description{
HierO is a tcltk based graphical user interface (GUI) for optimizing hierarchical data structure.
HierO constructs a user defined optimization problem to GAMS (General Algebraic Modeling System) form and uses Rneos package to send the problem to NEOS server for solving.
}
\details{
\tabular{ll}{
Package: \tab HierO\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-09-12\cr
License: \tab GPL-2\cr
Depends: \tab methods, rneos, RCurl, XMLRPC, bitops, XML, tcltk, tcltk2 \cr
}
}
\author{
Kari Tokola

Maintainer: Kari Tokola <kari.tokola@gmail.com>
}

\keyword{ package }


