\name{HierO}
\alias{HierO}

\title{GUI for hierarchical data}

\description{
HierO is a graphical user interface (GUI) tool for calculating optimal statistical power and sample size for hierarchical data structure.}
\usage{
HierO()
}

\details{
HierO constructs a user defined sample size optimization problem to GAMS (General Algebraic Modeling System) form and uses Rneos package to send the problem to NEOS server for solving.}

\author{
Kari Tokola
}

\references{\url{http://onlinelibrary.wiley.com/doi/10.1111/stan.12026/abstract}}
