\name{hieroClass-class}
\Rdversion{1.1}
\docType{class}
\alias{hieroClass-class}

\title{Class \code{"hieroClass"}}
\description{
A class for HierO objects
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{.S3Class}:}{Object of class \code{"character"}}
  }
}
\section{Methods}{
No methods defined with class "hieroClass" in the signature.
}

\author{
Kari Tokola
}

%% ~Make other sections like Warning with \section{Warning }{....} ~


\examples{
showClass("hieroClass")
}
\keyword{classes}
